+++ resources/views/dummy/block.blade.php
@@ -1,4 +1,4 @@
-<div class="w-100 border-dashed v-center"
+<div class="w-100 border-dashed v-center mt-3"
      style="min-height: 250px;">
     <h2 class="center font-thin label"
         style="color: #{{ random_int(0,7). random_int(3,6) . random_int(1,9) }}"